Edition des actions finalisées - gestion des articles creation et eddition

+++ database/migrations/2021_01_14_195441_create_actions_table.php
@@ -27,7 +27,7 @@ class CreateActionsTable extends Migration
                   ->on('members')
                   ->onDelete('cascade')
                   ->onUpdate('cascade');
-            $table->float('price')->nullable();
+            $table->decimal('price', 12, 2)->nullable();
         });
     }
